+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## Unreleased
 
+### Features
+- Refactor the nd2 source to use the nd2 library ([#797](../../pull/797))
+
 ### Improvements
 - Add options to threshold near min/max based on the histogram ([#798](../../pull/798))
 - Mark vsi extensions as being preferred by the bioformats source ([#793](../../pull/793))
+- Add mouse events to overlay annotations ([#794](../../pull/794))
 
 ## Version 1.11.2